pyproject.toml
[project]
name = "geo_search"
version = "0.1.0"
description = "geolocation agent"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.84.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
